fix: 충돌 해결을 위한 파일 및 컴포넌트 이름 변경
